按时间统计字段增加索引

+++ micro/dao.go
@@ -22,10 +22,10 @@ type Model struct {
 }
 
 type CountModel struct {
-	CountByYear  int8 `json:"count_by_year" sql:"index"`  // 按年统计
-	CountByMonth int8 `json:"count_by_month" sql:"index"` // 按月统计
-	CountByWeek  int8 `json:"count_by_week" sql:"index"`  // 按周统计
-	CountByDay   int8 `json:"count_by_day" sql:"index"`   // 按天统计
+	CountByYear  int16 `json:"count_by_year" sql:"index"`  // 按年统计
+	CountByMonth int16 `json:"count_by_month" sql:"index"` // 按月统计
+	CountByWeek  int16 `json:"count_by_week" sql:"index"`  // 按周统计
+	CountByDay   int16 `json:"count_by_day" sql:"index"`   // 按天统计
 }
 
 type Dao struct {
